requirements.txt
Flask==1.1.2
Flask-JWT==0.2.0
Flask-JWT-Extended==4.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
itsdangerous==2.0.1
Jinja2==3.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
faker
pika==1.3.1
psycopg2-binary
requests
pathlib==1.0.1
epaycosdk-win
flask-cors